chore: split models into files
